move pages/shared/editHelpers to lib/editors
