+++ lib/editors.js
@@ -1,11 +1,11 @@
 const ERR = require('async-stacktrace');
-const logger = require('../../lib/logger');
-const serverJobs = require('../../lib/server-jobs');
-const namedLocks = require('../../lib/named-locks');
-const syncFromDisk = require('../../sync/syncFromDisk');
-const courseUtil = require('../../lib/courseUtil');
-const requireFrontend = require('../../lib/require-frontend');
-const config = require('../../lib/config');
+const logger = require('../lib/logger');
+const serverJobs = require('../lib/server-jobs');
+const namedLocks = require('../lib/named-locks');
+const syncFromDisk = require('../sync/syncFromDisk');
+const courseUtil = require('../lib/courseUtil');
+const requireFrontend = require('../lib/require-frontend');
+const config = require('../lib/config');
 const path = require('path');
 const debug = require('debug')('prairielearn:' + path.basename(__filename, '.js'));
 const fs = require('fs-extra');
@@ -923,7 +923,7 @@ class QuestionAddEditor extends Editor {
                 });
             },
             (callback) => {
-                const fromPath = path.join(__dirname, '..', '..', 'exampleCourse', 'questions', 'demoCalculation');
+                const fromPath = path.join(__dirname, '..', 'exampleCourse', 'questions', 'demoCalculation');
                 const toPath = this.questionPath;
                 debug(`Copy template\n from ${fromPath}\n to ${toPath}`);
                 fs.copy(fromPath, toPath, {overwrite: false, errorOnExist: true}, (err) => {
