+++ lib/editors.sql
